requirements.txt
 langchain
 python-dotenv
 streamlit
 openai
 google-generativeai
 langchain-google-genai